+++ R/bind_trait_values.R
@@ -9,7 +9,7 @@
 #' @examples 
 #' \dontrun{
 #' traits <- austraits$traits %>% 
-#' dplyr::filter(dataset_id == "Falster_2005_1")
+#' dplyr::filter(dataset_id == "ABRS_1981")
 #' traits
 #' traits_bind <- bind_trait_values(traits)
 #' }
@@ -28,6 +28,7 @@ bind_trait_values <- function(trait_data) {
         .data %>% 
           dplyr::mutate(value = bind_x(.data$value),
                         value_type = bind_x(value_type),
+                        basis_of_value = bind_x(basis_of_value),
                         replicates = bind_x(replicates)) %>%
           dplyr::filter(dplyr::row_number()==1) 
       )
@@ -36,8 +37,8 @@ bind_trait_values <- function(trait_data) {
   }
   
   trait_data  %>% 
-    dplyr::group_by(observation_id, trait_name) %>% 
+    dplyr::group_by(dataset_id, observation_id, trait_name, method_id, method_context_id, repeat_measurements_id) %>% 
     bind_values_worker() %>% 
     dplyr::ungroup() %>% 
-    dplyr::arrange(observation_id, trait_name, value_type)
+    dplyr::arrange(dataset_id, observation_id, trait_name, value_type, method_id, method_context_id, repeat_measurements_id)
 }