+++ R/check_compatibility.R
@@ -17,7 +17,7 @@ check_compatibility <- function(austraits) {
     
     compatible <- FALSE
     
-    message("You are working with AusTraits version 3.0 or earlier. \nThis database structure is unsupported by the current version of this package. \nPlease see https://github.com/traitecoevo/austraits for details on installing old versions of the package.")
+    # message("You are working with AusTraits version 3.0 or earlier. \nThis database structure is unsupported by the current version of this package. \nPlease see https://github.com/traitecoevo/austraits for details on installing old versions of the package.")
     
   } else {
     
@@ -31,7 +31,7 @@ check_compatibility <- function(austraits) {
     } else{
       compatible <- FALSE
       
-      message("You are working with AusTraits version 4, which is unsupported by the current version of this package. \nPlease see https://github.com/traitecoevo/austraits for details on installing old versions of the package.")
+      # message("You are working with AusTraits version 4, which is unsupported by the current version of this package. \nPlease see https://github.com/traitecoevo/austraits for details on installing old versions of the package.")
     }
     
   }