+++ NEWS.md
@@ -1,2 +1,5 @@
-# austraits 2.2.1
+# austraits 3.0.0.9000
 - Updated dependencies, placing graphics related packages in Suggests
+- Added internal function to check the compatibility of databases
+- Added internal function notify users that some database versions will not be supported
+- No longer support AusTraits version <5.0.0 in `extract_*` functions